requirements.txt
python-dotenv==1.0.1
gmpy2==2.2.1
matplotlib==3.9.0
pandas==2.2.2
web3==6.19.0
loguru==0.7.0
wandb==0.18.7
annotated-types==0.7.0
anyio==4.6.2.post1
bittensor==9.0.1
bittensor-cli==9.0.1
colorama==0.4.6
decorator==5.1.1
fastapi==0.110.3
h11==0.14.0
msgpack==1.1.0
msgpack-numpy-opentensor==0.5.0
nest-asyncio==1.6.0
numpy==2.0.2
py==1.11.0
pydantic==2.10.0
pydantic-core==2.27.0
python-statemachine==2.4.0
retry==0.9.2
setuptools==70.0.0
sniffio==1.3.1
starlette==0.37.2
uvicorn==0.32.1
async-lru==2.0.5